Split YAML editor into partial files and extract its structs
- Carve SerializeReferenceYamlEditor into .Read/.Rewrite/.Scan partials, leaving the host file as the shared shell
- Pull the 5 nested types (ManagedTypeName, MissingReferenceEntry, RewriteEdit, RequiredFieldDescriptor, RequiredViolationEntry) into their own files
- Rename the editor assembly from SerializeReferences.Yaml to SerializeReferences and move AssemblyInfo/asmdef to the folder root
- Normalise brace/XML-doc style in the engine and add CheckNamespace banners to the test files
