+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Tests/Editor/SerializeReferences/Yaml/SerializeReferenceYamlEditorReadEdgeTests.cs
@@ -1,6 +1,7 @@
 using System.IO;
 using NUnit.Framework;
 
+// ReSharper disable once CheckNamespace
 namespace Aspid.FastTools.SerializeReferences.Editors.Tests
 {
     /// <summary>
