+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Aspid.FastTools.Unity.Editor.SerializeReferences.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "Aspid.FastTools.Unity.Editor.SerializeReferences.Yaml",
+    "name": "Aspid.FastTools.Unity.Editor.SerializeReferences",
     "rootNamespace": "",
     "references": [],
     "includePlatforms": [
