+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/ManagedTypeName.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4b8dd1b275f04c2586c1759a98fbf0ac
