+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/MissingReferenceEntry.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9e524d4356f1456ca8c9b24eeb836332
