+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/RequiredFieldDescriptor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e9e2863b6a414ef5a68b6838bf0a19f0
