+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/RequiredViolationEntry.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e4b58fe731a64c938472b9a20686c61f
