+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/RewriteEdit.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 033507a59ebb41f59333a6412b419d35
