+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/SerializeReferenceYamlEditor.Read.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1f04d1bb82d54cdfbb081cd9c36011f2
