+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/SerializeReferenceYamlEditor.Rewrite.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9dd83efa6e714171becc8d3092e030d7
