+++ Aspid.FastTools/Packages/tech.aspid.fasttools/Unity/Editor/Scripts/SerializeReferences/Yaml/SerializeReferenceYamlEditor.Scan.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f71f465426504d069f25e34332ceb514
